relocate CsvFileViewer
